feat(ConversationCells): set columns to have fixed width
