+++ src/script/components/CellsGlobalView/CellsTable/CellsTableNameColumn/CellsTableNameColumn.styles.ts
@@ -55,7 +55,6 @@ export const playIconStyles: CSSObject = {
 export const desktopNameStyles: CSSObject = {
   textOverflow: 'ellipsis',
   overflow: 'hidden',
-  // maxWidth: '300px',
   maxWidth: '100%',
   whiteSpace: 'nowrap',
   fontWeight: 'var(--font-weight-semibold)',
